+++ azalea/src/lib.rs
@@ -264,8 +264,18 @@ pub struct NoState;
 #[derive(Clone, Debug, Default)]
 #[non_exhaustive]
 pub struct JoinOpts {
-    /// The Socks5 proxy that this bot will use.
-    pub proxy: Option<Proxy>,
+    /// The SOCKS5 proxy that this bot will use for connecting to the Minecraft
+    /// server.
+    pub server_proxy: Option<Proxy>,
+    /// The SOCKS5 proxy that will be used when authenticating the bot's join
+    /// with Mojang.
+    ///
+    /// This should typically be either the same as [`Self::server_proxy`] or
+    /// `None`.
+    ///
+    /// This is useful to set if a server has `prevent-proxy-connections`
+    /// enabled.
+    pub sessionserver_proxy: Option<Proxy>,
     /// Override the server address that this specific bot will send in the
     /// handshake packet.
     pub custom_address: Option<ServerAddress>,
@@ -280,8 +290,11 @@ impl JoinOpts {
     }
 
     pub fn update(&mut self, other: &Self) {
-        if let Some(proxy) = other.proxy.clone() {
-            self.proxy = Some(proxy);
+        if let Some(proxy) = other.server_proxy.clone() {
+            self.server_proxy = Some(proxy);
+        }
+        if let Some(proxy) = other.sessionserver_proxy.clone() {
+            self.sessionserver_proxy = Some(proxy);
         }
         if let Some(custom_address) = other.custom_address.clone() {
             self.custom_address = Some(custom_address);
@@ -291,12 +304,38 @@ impl JoinOpts {
         }
     }
 
-    /// Set the proxy that this bot will use.
+    /// Configure the SOCKS5 proxy used for connecting to the server and for
+    /// authenticating with Mojang.
+    ///
+    /// To configure these separately, for example to only use the proxy for the
+    /// Minecraft server and not for authentication, you may use
+    /// [`Self::server_proxy`] and [`Self::sessionserver_proxy`] individually.
     #[must_use]
-    pub fn proxy(mut self, proxy: Proxy) -> Self {
-        self.proxy = Some(proxy);
+    pub fn proxy(self, proxy: Proxy) -> Self {
+        self.server_proxy(proxy.clone()).sessionserver_proxy(proxy)
+    }
+    /// Configure the SOCKS5 proxy that will be used for connecting to the
+    /// Minecraft server.
+    ///
+    /// To avoid errors on servers with the "prevent-proxy-connections" option
+    /// set, you should usually use [`Self::proxy`] instead.
+    ///
+    /// Also see [`Self::sessionserver_proxy`].
+    #[must_use]
+    pub fn server_proxy(mut self, proxy: Proxy) -> Self {
+        self.server_proxy = Some(proxy);
         self
     }
+    /// Configure the SOCKS5 proxy that this bot will use for authenticating the
+    /// server join with Mojang's API.
+    ///
+    /// Also see [`Self::proxy`] and [`Self::server_proxy`].
+    #[must_use]
+    pub fn sessionserver_proxy(mut self, proxy: Proxy) -> Self {
+        self.sessionserver_proxy = Some(proxy);
+        self
+    }
+
     /// Set the custom address that this bot will send in the handshake packet.
     #[must_use]
     pub fn custom_address(mut self, custom_address: ServerAddress) -> Self {